refactor: rename to eth_call
